[ios] Replace `LocationName` with `CoordinatePlacemarkData`
`CoordinatePlacemarkData` will be used to fetch `PlacemarkData` for a given coordinate to potentially avoid making calls with `CLGeocoder`.

We can't include `CLPlacemark` in the `@Model` because `CLPlacemark` is not Codable and can't be extended to be. Therefore we create a new `PlacemarkData` struct that includes the relevant fields that we need. We will create a `CLPlacemark` constructor for `PlacemarkData` to make things easier to work with.