[ios] Temporarily remove `Map` from `AimlessDriveListCell`
We want to see if we can find a repro for the issue @sridmad is facing. Before we can do that, we're trying to narrow down what might be causing the application to be killed by watchdog. The `Map` components in `AimlessDriveListCell` seem to use a decent amount of memory, so want to see if removing them makes any difference.