[ios] Fetch `currentUserID` only once and provide to hierarchy via `Environment`
We were previously fetching this record a stupid amount of times in the `onAppear` of `SocialFeedPostEngagement`.

While this got things working, we had two CloudKit queries that we were `await`ing sequentially. This is data that shouldn't really change and can be fetched just once. Might be some slight performance improvements, but things were fast enough before.