[ios] Introduce `fetchUserProfiles() async -> [UserProfile]`
So we can fetch these all in one go instead of on-demand for every single post, comment, etc.

We should also implement a version of this that takes in a set of userIDs so we're not fetching extraneous data, but there are so few users right now it doesn't really matter.